Extract config building into its own class
